Moving a form into Forms::Core
> A container for classes that are core to Sipity; That is to say
> things I don't believe to be part of a potential plugin.
>
> Why not Base? I believe Base to be more representative of things you
> build from, not stand alone elements.
