+++ app/repositories/sipity/queries/account_profile_queries.rb
@@ -3,7 +3,7 @@ module Sipity
     # Queries
     module AccountProfileQueries
       def build_account_profile_form(user:, attributes:)
-        Forms::ManageAccountProfileForm.new(user: user, repository: self, attributes: attributes)
+        Forms::Core::ManageAccountProfileForm.new(user: user, repository: self, attributes: attributes)
       end
     end
   end
